removed unnecessary cast
git-svn-id: https://svn.java.net/svn/skinlf~svn/trunk@238 184bcedb-ef5e-6bdb-aea7-fde92d8c4e55

+++ src/main/com/l2fprod/contrib/freehep/PanelArtistUtilities.java
@@ -72,7 +72,7 @@ public class PanelArtistUtilities {
    * @return        The YFlipTransform value
    */
   public static AffineTransform getYFlipTransform(int height) {
-    return new AffineTransform(1., 0., 0., -1., 0., (double) height);
+    return new AffineTransform(1., 0., 0., -1., 0., height);
   }
 
   /**
@@ -94,7 +94,7 @@ public class PanelArtistUtilities {
         ((double) width) / height,
         0.,
         0.,
-        (double) height);
+        height);
   }
 
   /**
